Module metric_engine::engine::options

Specific options for the metric engine to create or open a region.

Constants

IGNORE_COLUMN_IDS_FOR_DATA_REGION 🔒

Ignore building index on the column tsid which is unfriendly to the inverted index and will occupy excessive space if indexed.

SEG_ROW_COUNT_FOR_DATA_REGION 🔒

The empirical value for the seg row count of the metric data region. Compared to the mito engine, the pattern of the metric engine constructs smaller indices. Therefore, compared to the default seg row count of 1024, by adjusting it to a smaller value and appropriately increasing the size of the index, it results in an improved indexing effect.

Functions

set_index_options_for_data_region

Set the index options for the data region.

set_memtable_options_for_data_region

Set memtable options for the data region.