metric_engine::engine

Module options

Specific options for the metric engine to create or open a region.

Structs

PhysicalRegionOptions

Physical region options.

Enums

IndexOptions

Index options for auto created columns

Constants

SEG_ROW_COUNT_FOR_DATA_REGION ๐

The empirical value for the seg row count of the metric data region. Compared to the mito engine, the pattern of the metric engine constructs smaller indices. Therefore, compared to the default seg row count of 1024, by adjusting it to a smaller value and appropriately increasing the size of the index, it results in an improved indexing effect.

Functions

set_data_region_options

Sets data region specific options.